Update PLOTS_simplot.R
I extracted some plot validation work into a plot.data.validation function, which I then called from simplot(). Some of the parameters were modified in the extracted code, so the function returns a list() and that data needs to be extracted.

I also changed the remaining occurrences of 'both' to 'sim.and.data'